Function App/requirements.txt
azure-functions
requests
google-auth
